+++ ChangeLog.md
@@ -1,1 +1,4 @@
-+ Thu Apr 19 14:02:09 CST 2018: 将包名改为bixin_api
++ Thu Apr 19 CST 2018: 将包名改为bixin_api
+  + 将包名改为bixin_api
+  + 将`login/utils/login.mk_qr_code_cls` 重命名为 `mk_login_session_cls`
+  + 将`poim_user_id` 重命名为 `bixin_user_id`